# server/openapi.yaml
openapi: 3.1.0
info:
  title: MetaExtract API
  description: |
    MetaExtract is a comprehensive metadata extraction platform that extracts up to 45,000+
    metadata fields from any file. This API provides endpoints for file analysis, forensic
    investigation, and metadata management.

    ## Features
    - **Universal Metadata Extraction**: Extract metadata from images, videos, audio, documents, and scientific files
    - **Forensic Analysis**: Steganography detection, manipulation detection, AI content detection
    - **Timeline Reconstruction**: Chronological analysis of file events
    - **Batch Processing**: Process multiple files simultaneously (Professional+ tiers)

    ## Tiers
    | Tier | Fields | Max Size | File Types |
    |------|--------|----------|------------|
    | Free | ~200 | 10MB | Images |
    | Professional | ~1,000 | 100MB | Images + RAW |
    | Forensic | ~15,000 | 500MB | All media |
    | Enterprise | 45,000+ | 2GB | All files |

  version: 2.0.0
  contact:
    name: MetaExtract Support
    url: https://metaextract.io/support
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: /api
    description: API Base URL
...
paths:
...
  /forensic/capabilities:
    get:
      tags:
        - Forensic
      summary: Get forensic analysis capabilities
      description: |
        Returns available forensic analysis modules based on subscription tier.
      operationId: getForensicCapabilities
      parameters:
        - name: tier
          in: query
          schema:
            type: string
            enum: [free, professional, forensic, enterprise]
            default: enterprise
      responses:
        '200':
          description: Capabilities list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ForensicCapabilities'
  /compare/batch:
    post:
      tags:
        - Forensic
      summary: Compare metadata across multiple files
      description: |
        Upload 2 or more files to compare their metadata. Identifies matching fields,
        differences, and calculates similarity scores.
      operationId: batchCompare
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - files
              properties:
                files:
                  type: array
                  items:
                    type: string
                    format: binary
                  minItems: 2
                  maxItems: 50
      responses:
        '200':
          description: Comparison complete
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ComparisonResponse'
  /timeline/reconstruct:
    post:
      tags:
        - Forensic
      summary: Reconstruct timeline from multiple files
      description: |
        Analyze multiple files to reconstruct a chronological timeline of events.
        Identifies gaps and chain of custody issues.
      operationId: reconstructTimeline
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - files
              properties:
                files:
                  type: array
                  items:
                    type: string
                    format: binary
                  maxItems: 100
      responses:
        '200':
          description: Timeline reconstructed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TimelineResponse'
  /forensic/report:
    post:
      tags:
        - Forensic
      summary: Generate comprehensive forensic report
      description: |
        Generate a professional forensic report including all analysis results,
        conclusions, and recommendations. Enterprise tier only.
      operationId: generateForensicReport
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - files
              properties:
                files:
                  type: array
                  items:
                    type: string
                    format: binary
      responses:
        '200':
          description: Forensic report generated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ForensicReport'
...
components:
  schemas:
...
    ForensicCapabilities:
      type: object
      properties:
        tier:
          type: string
        advanced_analysis_available:
          type: boolean
        modules:
          type: object
          additionalProperties:
            type: object
            properties:
              available:
                type: boolean
              description:
                type: string
              methods:
                type: array
                items:
                  type: string
    ComparisonResponse:
      type: object
      properties:
        success:
          type: boolean
        comparison_id:
          type: string
          format: uuid
        files_compared:
          type: integer
        comparisons:
          type: array
          items:
            type: object
            properties:
              file1:
                type: string
              file2:
                type: string
              similarity_score:
                type: integer
              matching_fields:
                type: integer
              different_fields:
                type: integer
    TimelineResponse:
      type: object
      properties:
        success:
          type: boolean
        timeline_id:
          type: string
          format: uuid
        files_analyzed:
          type: integer
        events:
          type: array
          items:
            type: object
            properties:
              timestamp:
                type: string
              event_type:
                type: string
              source:
                type: string
              file:
                type: string
        gaps:
          type: array
          items:
            type: object
        chain_of_custody_complete:
          type: boolean
    ForensicReport:
      type: object
      properties:
        report_id:
          type: string
          format: uuid
        generated_at:
          type: string
          format: date-time
        analyst:
          type: string
        case_summary:
          type: object
        files:
          type: array
          items:
            type: object
        conclusions:
          type: object
        recommendations:
          type: array
          items:
            type: string
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    cookieAuth:
      type: apiKey
      in: cookie
      name: session
security:
  - bearerAuth: []
  - cookieAuth: []